requirements.txt
paho-mqtt==1.5.0
psutil==5.6.6
rpi_bad_power==0.1.0
pySMART==1.1.0
